player's waypoints avoid land
When player selects a waypoint on land, automatically selects the point
on the edge of land closest to that point.

Need to add code to produce intermediate waypoints to avoid crossing
over land. Possibly recursive by generating line segments and then
checking if they pass through land and so on.